added tests for events page view
