+++ securedpi_locks/tests.py
@@ -70,78 +70,6 @@ class SetupTestCase(TestCase):
             reverse('manual_unlock', kwargs={'pk': self.lock1.pk, 'action': 'unlock'}))
 
 
-class DashboardViewTestCase(SetupTestCase):
-    """Define test class for Dashboard view."""
-    def setUp(self):
-        """Define setup for tests."""
-        self.setUp = super(DashboardViewTestCase, self).setUp()
-        self.url = reverse('dashboard')
-        self.response = self.client.get(self.url)
-        self.template = 'securedpi/dashboard.html'
-
-    def test_auth_user_has_access_to_dashboard(self):
-        """Prove that response code is 200 for auth users."""
-        self.assertEquals(self.response.status_code, 200)
-
-    def test_right_template_is_used(self):
-        """Prove that right template is used to render page."""
-        self.assertTemplateUsed(self.response, self.template)
-
-    def test_locks_in_context(self):
-        """Prove that 'locks' are in response context."""
-        self.assertIn('locks', self.response.context)
-
-    def test_correct_number_of_locks_on_dashboard(self):
-        """Prove that correct number of locks renders on the dashboard."""
-        self.assertEqual(self.response.context['locks'].count(), 2)
-
-    def test_buttons_display_if_unlocked(self):
-        """
-        Make sure <Lock> button shows up and <Unlock> doesn't
-        if the lock.status == 'unlocked'."""
-        self.lock1.status = 'unlocked'
-        self.lock1.is_active = True
-        self.lock1.save()
-        response = self.client.get(self.url)
-        self.assertContains(response, self.expected1)
-        self.assertNotContains(response, self.expected2)
-
-
-    def test_buttons_display_if_locked(self):
-        """
-        Make sure <Unlock> button shows up and <Lock> doesn't
-        if the lock.status == 'locked'."""
-        self.lock1.status = 'locked'
-        self.lock1.is_active = True
-        self.lock1.save()
-        response = self.client.get(self.url)
-        self.assertNotContains(response, self.expected1)
-        self.assertContains(response, self.expected2)
-
-
-    def test_buttons_display_if_pending(self):
-        """
-        Make sure both <Unlock> and <Lock> buttons show up
-        if the lock.status == 'pending'."""
-        self.lock1.status = 'pending'
-        self.lock1.is_active = True
-        self.lock1.save()
-        response = self.client.get(self.url)
-        self.assertContains(response, self.expected1)
-        self.assertContains(response, self.expected2)
-
-
-    def test_buttons_display_if_not_active(self):
-        """
-        Make sure <Unlock> and <Lock> buttons don't show up
-        if the lock.is_active == False."""
-        self.lock1.is_active = False
-        self.lock1.save()
-        response = self.client.get(self.url)
-        self.assertNotContains(response, self.expected1)
-        self.assertNotContains(response, self.expected2)
-
-
 class EditLockTestCase(SetupTestCase):
     """Define test case class for lock editing."""
     def setUp(self):
